+++ src/index.js
@@ -1,9 +1,10 @@
-import * as content from "./content";
+import * as content from "./layouts/sample/content";
+import './index.css';
 
 content.render();
 
 if (module.hot) {
-  module.hot.accept('./content', () => {
+  module.hot.accept('./layouts/sample/content', () => {
     content.clean();
     content.render();
   });
