+++ src/layouts/sample/content-sample.js
